requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
python-dotenv
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.29.0
frozenlist==1.3.0
idna==3.3
ipython-genutils==0.2.0
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mistune==0.8.4
multidict==6.0.2
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
Pillow==9.0.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tabulate==0.8.9
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
webencodings==0.5.1
yarl==1.7.2
